requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15.2
    # via requests
charset-normalizer==2.1.1
    # via requests
google-api-core==2.10.0
    # via google-api-python-client
google-api-python-client==2.61.0
    # via -r requirements.in
google-auth==2.11.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==0.5.2
    # via -r requirements.in
googleapis-common-protos==1.56.4
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
numpy==1.23.3
    # via pandas
oauthlib==3.2.1
    # via requests-oauthlib
pandas==1.4.4
    # via -r requirements.in
protobuf==4.21.6
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via httplib2
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.2.1
    # via pandas
requests==2.28.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
tabulate==0.9.0
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
